--- Fabric.Identity.API/Models/ISoftDelete.cs
@@ -1,12 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace Fabric.Identity.API.Models
-{
-    public interface ISoftDelete
-    {
-        bool IsDeleted { get; set; }
-    }
-}
