+++ Fabric.Identity.API/Persistence/SqlServer/Models/ISoftDelete.cs
@@ -1,0 +1,7 @@
+﻿namespace Fabric.Identity.API.Persistence.SqlServer.Models
+{
+    public interface ISoftDelete
+    {
+        bool IsDeleted { get; set; }
+    }
+}
