+++ Fabric.Identity.API/Persistence/SqlServer/Stores/SqlServerClientStore.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
-using Fabric.Identity.API.Models.Domain;
+using Fabric.Identity.API.Persistence.SqlServer.Models;
 using Fabric.Identity.API.Persistence.SqlServer.Services;
 using IdentityServer4.EntityFramework.Mappers;
 using IdentityServer4.Models;
